Move read button functionality to App from Books
