+++ src/Components/App/App.js
@@ -7,7 +7,7 @@ import BookInfo from '../BookInfo/BookInfo'
 import ReadingList from '../ReadingList/ReadingList'
 import { fetchBooks } from '../../API';
 import { connect } from 'react-redux'
-import { setBooks, setList } from '../../actions';
+import { setBooks, setList, addFavorite } from '../../actions';
 import { bindActionCreators } from 'redux';
 import { Route, Switch } from 'react-router-dom';
 
@@ -33,7 +33,6 @@ class App extends Component {
         const response = await fetchBooks(url);
         setBooks(response.results.books);
         const books = response.results.books;
-        console.log('books after setBooks', books)
         return setList(url, books.map(book => book.primary_isbn10));
       })
     }
@@ -45,7 +44,6 @@ class App extends Component {
   filterBooks = (listName) => {
     const listOfIds = this.props.lists[listName]
     const filteredBooks = this.props.books.filter(book => listOfIds.includes(book.primary_isbn10))
-    console.log('filtered books from filterBooks', filteredBooks)
     return filteredBooks;
   }
 
@@ -54,10 +52,9 @@ class App extends Component {
     const listsKeys = Object.keys(listUrls);
     const listBooks = listsKeys.map(listName => {
       const filteredBooks = this.filterBooks(listName)
-      console.log('filtered books from createBookLists', filteredBooks)
       if(filteredBooks.length > 0) {
         return (
-          <Books key={listName} id={listName} listName={listName} filteredBooks={filteredBooks}/>
+          <Books key={listName} id={listName} listName={listName} filteredBooks={filteredBooks} addBook={this.handleClick} />
         )
       }
     })
@@ -70,8 +67,33 @@ class App extends Component {
         this.setState({ foundBooks: [book] })
       }
     })
-    console.log(this.state.foundBooks)
     return findBooks;
+  }
+
+  changeButtonStyling(id) {
+    const allButtons = document.querySelectorAll('.add-to-reading-list')
+    const foundButton = Array.from(allButtons).find(button => button.id === id)
+    if (foundButton.classList.contains('active')) {
+      foundButton.classList.remove('active')
+      foundButton.classList.add('inactive')
+    } else if (foundButton.classList.contains('inactive')) {
+      foundButton.classList.remove('inactive')
+      foundButton.classList.add('active')
+    }
+  }
+
+  handleClick = (event) => {
+    const { addFavorite, readingList, books } = this.props;
+    const id = event.target.id;
+    const foundBook = books.find(book => book.primary_isbn10 === id);
+    const isOnList = readingList.includes(foundBook)
+
+    if (!isOnList) {
+      addFavorite(foundBook) 
+      this.changeButtonStyling(id)
+    } else {
+      this.changeButtonStyling(id)
+    }
   }
 
   render() {
@@ -105,7 +127,7 @@ class App extends Component {
             } />
             <Route exact path='/:bookId' render={({ match }) => {
               const bookClicked = books.find((book) => book.primary_isbn10 == parseInt(match.params.bookId))
-              return <BookInfo book={bookClicked} /> }}
+              return <BookInfo book={bookClicked} addBook={this.handleClick} /> }}
             />
           </Switch>
       </div>
@@ -122,7 +144,8 @@ export const mapStateToProps = ({ books, readingList, lists }) => ({
 export const mapDispatchToProps = dispatch => (
   bindActionCreators({
     setBooks,
-    setList
+    setList,
+    addFavorite
   }, dispatch)
 )
